requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1